+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [7.8.0] - 2026-09-23
+
+_Stable release based on [7.8.0-rc.1]._
+
 ## [7.8.0-rc.1] - 2026-09-23
 
 ### Added
@@ -2073,6 +2077,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[7.8.0]: https://https://github.com/internetguru/laravel-common/compare/v7.7.3...v7.8.0
 [7.8.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v7.7.3
 [7.7.3]: https://https://github.com/internetguru/laravel-common/compare/v7.7.2...v7.7.3
 [7.7.2]: https://https://github.com/internetguru/laravel-common/compare/v7.7.1...v7.7.2